% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cetsa.R
\name{read_cetsa}
\alias{read_cetsa}
\title{Read in protein data from proteome discoverer}
\usage{
read_cetsa(
  protein_path,
  peptide_path,
  Frac = TRUE,
  temps = set_temps(11, c(37, 41, 44, 47, 50, 53, 56, 59, 63, 67, 68)),
  solvent = "DMSO",
  CARRIER = TRUE
)
}
\arguments{
\item{f.}{Path of Excel file output from Proteome Discoverer}
}
\value{
a dataframe containing extracted information
}
\description{
Read in Excel file and apply minimal pre-processing
}
